restructured model structure
