+++ mix.exs
@@ -34,6 +34,7 @@ defmodule GenDSL.MixProject do
       # {:dep_from_hexpm, "~> 0.3.0"},
       # {:dep_from_git, git: "https://github.com/elixir-lang/my_dep.git", tag: "0.1.0"}
       {:ex_doc, "~> 0.27", only: :dev, runtime: false},
+      {:stream_data, "~> 0.6.0", only: :test},
       {:poison, "~> 5.0"},
       {:exconstructor, "~> 1.2"},
       {:ecto, "~> 3.9"}
