+++ test/gen_dsl_test.exs
@@ -5,4 +5,9 @@ defmodule GenDSLTest do
   test "greets the world" do
     assert GenDSL.hello() == :world
   end
+
+  test "HTML elements" do
+  end
+
+  test ""
 end
